backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-docx
openai